fix: changed all routing from profile to hackathon page
